refactor(#104): extract blossom-core lib crate for native test execution
Move pure Rust types (BlobStatus, BlobMetadata, BlossomError) and
policy functions (validate_sha256_format, map_moderate_action,
handle_creator_delete_with_ops, build_creator_delete_response) into a
blossom-core workspace member with no Fastly SDK dependency.

The main binary crate re-exports everything via pub use, so all
existing crate::blossom::* and crate::error::* imports keep working.
Fastly-dependent functions (soft_delete_blob, restore_soft_deleted_blob,
DefaultCreatorDeleteOps, handle_creator_delete) stay in the binary
crate.

cargo test -p blossom-core now runs 88 tests natively (0.00s) without
Viceroy. The compiled WASM binary is unchanged.

Relates to divinevideo/divine-blossom#104
