+++ blossom-core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "blossom-core"
+version = "0.1.0"
+edition = "2021"
+description = "Pure-logic types and policy functions for divine-blossom, testable without Viceroy"
+
+[dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+http = "1"
